+++ tests/testthat/test_fhir_collapse.R
@@ -10,7 +10,7 @@ testthat::test_that(
 						verbose = 0
 		)
 
-		fhir_collapse(d, columns = "address.line", sep = "|", brackets = brackets)
+		fhir_collapse(d, columns = "address.line", sep = "|", brackets = c("[", "]"))
 		}, style = "json2")
 
 	})
@@ -22,11 +22,11 @@ testthat::test_that(
 		d <- fhir_crack(bundles,
 						design = fhir_table_description(
 							resource = "Patient",
-							brackets = brackets,
-							sep = sep),
+							brackets = c("[", "]"),
+							sep = "|"),
 						verbose = 0
 		)
-		fhir_collapse(d, columns = "name.given", sep = sep, brackets = brackets)
+		fhir_collapse(d, columns = "name.given", sep = "|", brackets = c("[", "]"))
 		}, style = "json2")
 
 	})